# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  clientName: string;
  clientEmail: string;
  name: string;
  type: string;
  status: string;
  budget: number;
  deadline: string;
  progress: number;
  accent: string;
  initials: string;
  shareToken?: string;
}

export interface ProjectInput {
  clientName: string;
  clientEmail: string;
  name: string;
  type: string;
  goals: string;
  budget: number;
  deadline: string;
  notes?: string;
}

export interface ProjectUpdate {
  name?: string;
  clientName?: string;
  clientEmail?: string;
  type?: string;
  status?: string;
  progress?: number;
  budget?: number;
  deadline?: string;
  goals?: string;
  notes?: string;
}

export interface Proposal {
  status: string;
  headline: string;
  body: string;
  /** @nullable */
  selectedPackage: string | null;
}

export interface Package {
  id: string;
  name: string;
  price: number;
  description: string;
  features: string[];
  recommended: boolean;
}

export interface Milestone {
  id: string;
  name: string;
  date: string;
  status: string;
}

export interface Task {
  id: string;
  title: string;
  phase: string;
  status: string;
  dueDate: string;
  assignee?: string;
}

export interface Invoice {
  id: string;
  number: string;
  amount: number;
  dueDate: string;
  status: string;
  description: string;
}

export interface Activity {
  id: string;
  actor: string;
  action: string;
  time: string;
  type: string;
}

export type ProjectDetail = Project & {
  goals: string;
  notes: string;
  proposal: Proposal;
  packages: Package[];
  milestones: Milestone[];
  tasks: Task[];
  invoices: Invoice[];
  activities: Activity[];
};

export interface ProposalUpdate {
  status?: string;
  headline?: string;
  body?: string;
  /** @nullable */
  selectedPackage?: string | null;
}

export interface MilestoneInput {
  name: string;
  date: string;
  status: string;
}

export interface MilestoneUpdate {
  name?: string;
  date?: string;
  status?: string;
}

export interface TaskUpdate {
  status?: string;
  title?: string;
  phase?: string;
  dueDate?: string;
  assignee?: string;
}

export interface TaskInput {
  title: string;
  phase: string;
  status: string;
  dueDate: string;
  assignee: string;
}

export interface InvoiceUpdate {
  status?: string;
  description?: string;
  amount?: number;
  dueDate?: string;
  number?: string;
}

export interface InvoiceInput {
  description: string;
  amount: number;
  dueDate: string;
  status: string;
  number?: string;
}

export type JobProposalInputTone = typeof JobProposalInputTone[keyof typeof JobProposalInputTone];


export const JobProposalInputTone = {
  confident: 'confident',
  consultative: 'consultative',
  warm: 'warm',
} as const;

export type JobProposalInputLength = typeof JobProposalInputLength[keyof typeof JobProposalInputLength];


export const JobProposalInputLength = {
  short: 'short',
  standard: 'standard',
  detailed: 'detailed',
} as const;

export type JobProposalInputProfile = {
  name?: string;
  studio?: string;
  email?: string;
  bio?: string;
  portfolio?: string;
};

export interface JobProposalInput {
  url?: string;
  description?: string;
  tone?: JobProposalInputTone;
  length?: JobProposalInputLength;
  profile?: JobProposalInputProfile;
}

export type JobProposalJob = {
  title?: string;
  org?: string;
  url?: string;
  description?: string;
  employmentType?: string;
};

export interface ProposalDraft {
  id: string;
  proposal: string;
  source: string;
  createdAt: string;
  title?: string;
  org?: string;
  url?: string;
  tone?: string;
  length?: string;
}

export interface JobProposal {
  proposal: string;
  source: string;
  needsDescription?: boolean;
  draft?: ProposalDraft;
  job?: JobProposalJob;
}

export interface ProposalDecision {
  packageId?: string;
  note?: string;
}

export interface Upcoming {
  projectId: string;
  projectName: string;
  clientName: string;
  label: string;
  date: string;
  daysLeft: number;
}

export interface ChartPoint {
  month: string;
  value: number;
}

export interface Dashboard {
  revenue: number;
  revenueChange: number;
  activeProjects: number;
  outstanding: number;
  outstandingCount: number;
  upcoming: Upcoming[];
  activity: Activity[];
  chart: ChartPoint[];
  projects: Project[];
}

export type GetProjectsParams = {
search?: string;
status?: string;
};