# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /job-proposals/generate:
    post:
      operationId: generateJobProposal
      tags: [workflow]
      summary: Draft a proposal for a job posting from a URL or pasted description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobProposalInput"
      responses:
        "200":
          description: Drafted proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProposal"
...
  /job-proposals/{id}:
    delete:
      operationId: deleteJobProposal
      tags: [workflow]
      summary: Delete a saved proposal draft
      parameters:
        - $ref: "#/components/parameters/ProposalDraftId"
      responses:
        "204":
          description: Draft deleted
...
components:
  parameters:
...
    ProposalDraftId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    JobProposalInput:
      type: object
      properties:
        url: { type: string }
        description: { type: string }
        tone: { type: string, enum: [confident, consultative, warm] }
        length: { type: string, enum: [short, standard, detailed] }
        profile:
          type: object
          properties:
            name: { type: string }
            studio: { type: string }
            email: { type: string }
            bio: { type: string }
            portfolio: { type: string }
    JobProposal:
      type: object
      required: [proposal, source]
      properties:
        proposal: { type: string }
        source: { type: string }
        needsDescription: { type: boolean }
        draft:
          $ref: "#/components/schemas/ProposalDraft"
        job:
          type: object
          properties:
            title: { type: string }
            org: { type: string }
            url: { type: string }
            description: { type: string }
            employmentType: { type: string }
    ProposalDraft:
      type: object
      required: [id, proposal, source, createdAt]
      properties:
        id: { type: string }
        proposal: { type: string }
        source: { type: string }
        createdAt: { type: string }
        title: { type: string }
        org: { type: string }
        url: { type: string }
        tone: { type: string }
        length: { type: string }
...