# lib/api-zod/src/generated/types/jobProposal.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobProposalJob } from './jobProposalJob';
import type { ProposalDraft } from './proposalDraft';

export interface JobProposal {
  proposal: string;
  source: string;
  needsDescription?: boolean;
  draft?: ProposalDraft;
  job?: JobProposalJob;
}