# lib/api-zod/src/generated/types/jobProposalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobProposalInputLength } from './jobProposalInputLength';
import type { JobProposalInputProfile } from './jobProposalInputProfile';
import type { JobProposalInputTone } from './jobProposalInputTone';

export interface JobProposalInput {
  url?: string;
  description?: string;
  tone?: JobProposalInputTone;
  length?: JobProposalInputLength;
  profile?: JobProposalInputProfile;
}