# lib/api-zod/src/generated/types/proposalDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalDraft {
  id: string;
  proposal: string;
  source: string;
  createdAt: string;
  title?: string;
  org?: string;
  url?: string;
  tone?: string;
  length?: string;
}